minor fixes

+++ controllers/postController.js
@@ -13,7 +13,7 @@ exports.index = async (req, res, next) => {
 exports.show = async (req, res, next) => {
   try {
     validationHandler(req);
-    const post = await Post.findOne({ _id: req.params.id });
+    const post = await Post.findOne({ _id: req.params.id }).populate("user");
     res.send(post);
   } catch (err) {
     next(err);
@@ -27,6 +27,7 @@ exports.store = async (req, res, next) => {
 
     post.description = req.body.description;
     post.image = req.file.filename;
+    post.userId = req.user._id;
 
     post = await post.save();
     res.send(post);
@@ -40,9 +41,15 @@ exports.update = async (req, res, next) => {
     validationHandler(req);
     let post = await Post.findById(req.params.id);
 
-    post.description = req.body.description;
+    if (!post || post.userId !== req.user._id) {
+      const error = new Error("Wrong request");
+      error.statusCode = 400;
+      throw error;
+    }
 
+    post.description = req.body.description;
     post = await post.save();
+
     res.send(post);
   } catch (err) {
     next(err);
@@ -54,8 +61,13 @@ exports.delete = async (req, res, next) => {
     validationHandler(req);
     let post = await Post.findById(req.params.id);
 
-    await post.delete();
+    if (!post || post.userId !== req.user._id) {
+      const error = new Error("Wrong request");
+      error.statusCode = 400;
+      throw error;
+    }
 
+    await post.delete();
     res.send({
       message: "Post deleted",
     });
